requirements.txt
streamlit>=1.40.0
faster-whisper>=1.1.1
openai>=1.62.0
elevenlabs>=1.19.0
deepl>=1.22.0
pydub>=0.25.1
audioop-lts>=0.2.1
httpx>=0.28.1
python-dotenv>=1.0.1
psycopg2-binary>=2.9.9 